requirements.txt
streamlit
sounddevice
numpy
SpeechRecognition
gtts
google-generativeai